safari-600.7-branch/LayoutTests/fast/css/counters/deep-before-expected.txt

PASS window.internals.counterValue(document.getElementById('div1')) is '0'

TEST COMPLETE
